Extract build_document_records, nest_flat_dict, and related methods to import_export.py
